requirements.txt
Django==4.2
phonenumbers
pytest
pytest-django
django-phonenumber-field
faker